Extract DS.ActiveModelSerializer embedded record support into mixin
