+++ packages/activemodel-adapter/tests/integration/active_model_serializer_test.js
@@ -1,5 +1,5 @@
 var get = Ember.get, set = Ember.set;
-var HomePlanet, league, SuperVillain, superVillain, EvilMinion, YellowMinion, DoomsdayDevice, PopularVillain, Comment, Course, Unit, env;
+var HomePlanet, league, SuperVillain, superVillain, EvilMinion, YellowMinion, DoomsdayDevice, MediocreVillain, env;
 
 module("integration/active_model - ActiveModelSerializer", {
   setup: function() {
@@ -22,22 +22,9 @@ module("integration/active_model - ActiveModelSerializer", {
       name:         DS.attr('string'),
       evilMinion:   DS.belongsTo('evilMinion', {polymorphic: true})
     });
-    PopularVillain = DS.Model.extend({
+    MediocreVillain = DS.Model.extend({
       name:         DS.attr('string'),
       evilMinions:  DS.hasMany('evilMinion', {polymorphic: true})
-    });
-    Comment = DS.Model.extend({
-      body: DS.attr('string'),
-      root: DS.attr('boolean'),
-      children: DS.hasMany('comment')
-    });
-    Course = DS.Model.extend({
-      name: DS.attr('string'),
-      prerequisiteUnits: DS.hasMany('unit'),
-      units: DS.hasMany('unit')
-    });
-    Unit = DS.Model.extend({
-      name: DS.attr('string')
     });
     env = setupStore({
       superVillain:   SuperVillain,
@@ -45,20 +32,14 @@ module("integration/active_model - ActiveModelSerializer", {
       evilMinion:     EvilMinion,
       yellowMinion:   YellowMinion,
       doomsdayDevice: DoomsdayDevice,
-      popularVillain: PopularVillain,
-      comment:        Comment,
-      course:         Course,
-      unit:           Unit
+      mediocreVillain: MediocreVillain
     });
     env.store.modelFor('superVillain');
     env.store.modelFor('homePlanet');
     env.store.modelFor('evilMinion');
     env.store.modelFor('yellowMinion');
     env.store.modelFor('doomsdayDevice');
-    env.store.modelFor('popularVillain');
-    env.store.modelFor('comment');
-    env.store.modelFor('course');
-    env.store.modelFor('unit');
+    env.store.modelFor('mediocreVillain');
     env.container.register('serializer:application', DS.ActiveModelSerializer);
     env.container.register('serializer:ams', DS.ActiveModelSerializer);
     env.container.register('adapter:ams', DS.ActiveModelAdapter);
@@ -132,215 +113,6 @@ test("extractSingle", function() {
   }));
 });
 
-test("extractSingle with embedded objects", function() {
-  env.container.register('adapter:superVillain', DS.ActiveModelAdapter);
-  env.container.register('serializer:homePlanet', DS.ActiveModelSerializer.extend({
-    attrs: {
-      villains: {embedded: 'always'}
-    }
-  }));
-
-  var serializer = env.container.lookup("serializer:homePlanet");
-  var json_hash = {
-    home_planet: {
-      id: "1",
-      name: "Umber",
-      villains: [{
-        id: "1",
-        first_name: "Tom",
-        last_name: "Dale"
-      }]
-    }
-  };
-
-  var json = serializer.extractSingle(env.store, HomePlanet, json_hash);
-
-  deepEqual(json, {
-    id: "1",
-    name: "Umber",
-    villains: ["1"]
-  });
-  env.store.find("superVillain", 1).then(async(function(minion) {
-    equal(minion.get('firstName'), "Tom");
-  }));
-});
-
-test("extractSingle with embedded objects inside embedded objects", function() {
-  env.container.register('adapter:superVillain', DS.ActiveModelAdapter);
-  env.container.register('serializer:homePlanet', DS.ActiveModelSerializer.extend({
-    attrs: {
-      villains: {embedded: 'always'}
-    }
-  }));
-  env.container.register('serializer:superVillain', DS.ActiveModelSerializer.extend({
-    attrs: {
-      evilMinions: {embedded: 'always'}
-    }
-  }));
-
-  var serializer = env.container.lookup("serializer:homePlanet");
-  var json_hash = {
-    home_planet: {
-      id: "1",
-      name: "Umber",
-      villains: [{
-        id: "1",
-        first_name: "Tom",
-        last_name: "Dale",
-        evil_minions: [{
-          id: "1",
-          name: "Alex"
-        }]
-      }]
-    }
-  };
-
-  var json = serializer.extractSingle(env.store, HomePlanet, json_hash);
-
-  deepEqual(json, {
-    id: "1",
-    name: "Umber",
-    villains: ["1"]
-  });
-  env.store.find("superVillain", 1).then(async(function(villain) {
-    equal(villain.get('firstName'), "Tom");
-    equal(villain.get('evilMinions.length'), 1, "Should load the embedded child");
-    equal(villain.get('evilMinions.firstObject.name'), "Alex", "Should load the embedded child");
-  }));
-  env.store.find("evilMinion", 1).then(async(function(minion) {
-    equal(minion.get('name'), "Alex");
-  }));
-});
-
-test("extractSingle with embedded objects of same type", function() {
-  env.container.register('adapter:comment', DS.ActiveModelAdapter);
-  env.container.register('serializer:comment', DS.ActiveModelSerializer.extend({
-    attrs: {
-      children: {embedded: 'always'}
-    }
-  }));
-
-  var serializer = env.container.lookup("serializer:comment");
-  var json_hash = {
-    comment: {
-      id: "1",
-      body: "Hello",
-      root: true,
-      children: [{
-        id: "2",
-        body: "World",
-        root: false
-      },
-      {
-        id: "3",
-        body: "Foo",
-        root: false
-      }]
-    }
-  };
-  var json = serializer.extractSingle(env.store, Comment, json_hash);
-
-  deepEqual(json, {
-    id: "1",
-    body: "Hello",
-    root: true,
-    children: ["2", "3"]
-  }, "Primary record was correct");
-  equal(env.store.recordForId("comment", "2").get("body"), "World", "Secondary records found in the store");
-  equal(env.store.recordForId("comment", "3").get("body"), "Foo", "Secondary records found in the store");
-});
-
-test("extractSingle with embedded objects inside embedded objects of same type", function() {
-  env.container.register('adapter:comment', DS.ActiveModelAdapter);
-  env.container.register('serializer:comment', DS.ActiveModelSerializer.extend({
-    attrs: {
-      children: {embedded: 'always'}
-    }
-  }));
-
-  var serializer = env.container.lookup("serializer:comment");
-  var json_hash = {
-    comment: {
-      id: "1",
-      body: "Hello",
-      root: true,
-      children: [{
-        id: "2",
-        body: "World",
-        root: false,
-        children: [{
-          id: "4",
-          body: "Another",
-          root: false
-        }]
-      },
-      {
-        id: "3",
-        body: "Foo",
-        root: false
-      }]
-    }
-  };
-  var json = serializer.extractSingle(env.store, Comment, json_hash);
-
-  deepEqual(json, {
-    id: "1",
-    body: "Hello",
-    root: true,
-    children: ["2", "3"]
-  }, "Primary record was correct");
-  equal(env.store.recordForId("comment", "2").get("body"), "World", "Secondary records found in the store");
-  equal(env.store.recordForId("comment", "3").get("body"), "Foo", "Secondary records found in the store");
-  equal(env.store.recordForId("comment", "4").get("body"), "Another", "Secondary records found in the store");
-  equal(env.store.recordForId("comment", "2").get("children.length"), 1, "Should have one embedded record");
-  equal(env.store.recordForId("comment", "2").get("children.firstObject.body"), "Another", "Should have one embedded record");
-});
-
-test("extractSingle with embedded objects of same type, but from separate attributes", function() {
-  env.container.register('adapter:course', DS.ActiveModelAdapter);
-  env.container.register('serializer:course', DS.ActiveModelSerializer.extend({
-    attrs: {
-      prerequisiteUnits: {embedded: 'always'},
-      units: {embedded: 'always'}
-    }
-  }));
-
-  var serializer = env.container.lookup("serializer:course");
-  var json_hash = {
-    course: {
-      id: "1",
-      name: "Course 1",
-      prerequisite_units: [{
-        id: "1",
-        name: "Unit 1"
-      },{
-        id: "3",
-        name: "Unit 3"
-      }],
-      units: [{
-        id: "2",
-        name: "Unit 2"
-      },{
-        id: "4",
-        name: "Unit 4"
-      }]
-    }
-  };
-  var json = serializer.extractSingle(env.store, Course, json_hash);
-
-  deepEqual(json, {
-    id: "1",
-    name: "Course 1",
-    prerequisiteUnits: ["1", "3"],
-    units: ["2", "4"]
-  }, "Primary array was correct");
-
-  equal(env.store.recordForId("unit", "1").get("name"), "Unit 1", "Secondary records found in the store");
-  equal(env.store.recordForId("unit", "2").get("name"), "Unit 2", "Secondary records found in the store");
-  equal(env.store.recordForId("unit", "3").get("name"), "Unit 3", "Secondary records found in the store");
-  equal(env.store.recordForId("unit", "4").get("name"), "Unit 4", "Secondary records found in the store");
-});
-
 test("extractArray", function() {
   env.container.register('adapter:superVillain', DS.ActiveModelAdapter);
 
@@ -362,151 +134,6 @@ test("extractArray", function() {
   }));
 });
 
-test("extractArray with embedded objects", function() {
-  env.container.register('adapter:superVillain', DS.ActiveModelAdapter);
-  env.container.register('serializer:homePlanet', DS.ActiveModelSerializer.extend({
-    attrs: {
-      villains: {embedded: 'always'}
-    }
-  }));
-
-  var serializer = env.container.lookup("serializer:homePlanet");
-
-  var json_hash = {
-    home_planets: [{
-      id: "1",
-      name: "Umber",
-      villains: [{
-        id: "1",
-        first_name: "Tom",
-        last_name: "Dale"
-      }]
-    }]
-  };
-
-  var array = serializer.extractArray(env.store, HomePlanet, json_hash);
-
-  deepEqual(array, [{
-    id: "1",
-    name: "Umber",
-    villains: ["1"]
-  }]);
-
-  env.store.find("superVillain", 1).then(async(function(minion){
-    equal(minion.get('firstName'), "Tom");
-  }));
-});
-
-test("extractArray with embedded objects of same type as primary type", function() {
-  env.container.register('adapter:comment', DS.ActiveModelAdapter);
-  env.container.register('serializer:comment', DS.ActiveModelSerializer.extend({
-    attrs: {
-      children: {embedded: 'always'}
-    }
-  }));
-
-  var serializer = env.container.lookup("serializer:comment");
-
-  var json_hash = {
-    comments: [{
-      id: "1",
-      body: "Hello",
-      root: true,
-      children: [{
-        id: "2",
-        body: "World",
-        root: false
-      },
-      {
-        id: "3",
-        body: "Foo",
-        root: false
-      }]
-    }]
-  };
-
-  var array = serializer.extractArray(env.store, Comment, json_hash);
-
-  deepEqual(array, [{
-    id: "1",
-    body: "Hello",
-    root: true,
-    children: ["2", "3"]
-  }], "Primary array is correct");
-
-  equal(env.store.recordForId("comment", "2").get("body"), "World", "Secondary record found in the store");
-  equal(env.store.recordForId("comment", "3").get("body"), "Foo", "Secondary record found in the store");
-});
-
-test("extractArray with embedded objects of same type, but from separate attributes", function() {
-  env.container.register('adapter:course', DS.ActiveModelAdapter);
-  env.container.register('serializer:course', DS.ActiveModelSerializer.extend({
-    attrs: {
-      prerequisiteUnits: {embedded: 'always'},
-      units: {embedded: 'always'}
-    }
-  }));
-
-  var serializer = env.container.lookup("serializer:course");
-  var json_hash = {
-    courses: [{
-      id: "1",
-      name: "Course 1",
-      prerequisite_units: [{
-        id: "1",
-        name: "Unit 1"
-      },{
-        id: "3",
-        name: "Unit 3"
-      }],
-      units: [{
-        id: "2",
-        name: "Unit 2"
-      },{
-        id: "4",
-        name: "Unit 4"
-      }]
-    },{
-      id: "2",
-      name: "Course 2",
-      prerequisite_units: [{
-        id: "1",
-        name: "Unit 1"
-      },{
-        id: "3",
-        name: "Unit 3"
-      }],
-      units: [{
-        id: "5",
-        name: "Unit 5"
-      },{
-        id: "6",
-        name: "Unit 6"
-      }]
-    }]
-  };
-  var json = serializer.extractArray(env.store, Course, json_hash);
-
-  deepEqual(json, [{
-    id: "1",
-    name: "Course 1",
-    prerequisiteUnits: ["1", "3"],
-    units: ["2", "4"]
-  },{
-    id: "2",
-    name: "Course 2",
-    prerequisiteUnits: ["1", "3"],
-    units: ["5", "6"]
-  }], "Primary array was correct");
-
-  equal(env.store.recordForId("unit", "1").get("name"), "Unit 1", "Secondary records found in the store");
-  equal(env.store.recordForId("unit", "2").get("name"), "Unit 2", "Secondary records found in the store");
-  equal(env.store.recordForId("unit", "3").get("name"), "Unit 3", "Secondary records found in the store");
-  equal(env.store.recordForId("unit", "4").get("name"), "Unit 4", "Secondary records found in the store");
-  equal(env.store.recordForId("unit", "5").get("name"), "Unit 5", "Secondary records found in the store");
-  equal(env.store.recordForId("unit", "6").get("name"), "Unit 6", "Secondary records found in the store");
-});
-
 test("serialize polymorphic", function() {
   var tom = env.store.createRecord(YellowMinion,   {name: "Alex", id: "124"});
   var ray = env.store.createRecord(DoomsdayDevice, {evilMinion: tom, name: "DeathRay"});
@@ -520,41 +147,17 @@ test("serialize polymorphic", function() {
   });
 });
 
-test("serialize with embedded objects", function() {
-  league = env.store.createRecord(HomePlanet, { name: "Villain League", id: "123" });
-  var tom = env.store.createRecord(SuperVillain, { firstName: "Tom", lastName: "Dale", homePlanet: league });
-
-  env.container.register('serializer:homePlanet', DS.ActiveModelSerializer.extend({
-    attrs: {
-      villains: {embedded: 'always'}
-    }
-  }));
-  var serializer = env.container.lookup("serializer:homePlanet");
-
-  var json = serializer.serialize(league);
-
-  deepEqual(json, {
-    name: "Villain League",
-    villains: [{
-      id: get(tom, "id"),
-      first_name: "Tom",
-      last_name: "Dale",
-      home_planet_id: get(league, "id")
-    }]
-  });
-});
-
 test("extractPolymorphic hasMany", function() {
   env.container.register('adapter:yellowMinion', DS.ActiveModelAdapter);
-  PopularVillain.toString   = function() { return "PopularVillain"; };
+  MediocreVillain.toString   = function() { return "MediocreVillain"; };
   YellowMinion.toString = function() { return "YellowMinion"; };
 
   var json_hash = {
-    popular_villain: {id: 1, name: "Dr Horrible", evil_minions: [{ type: "yellow_minion", id: 12}] },
+    mediocre_villain: {id: 1, name: "Dr Horrible", evil_minions: [{ type: "yellow_minion", id: 12}] },
     evil_minions:    [{id: 12, name: "Alex", doomsday_device_ids: [1] }]
   };
 
-  var json = env.amsSerializer.extractSingle(env.store, PopularVillain, json_hash);
+  var json = env.amsSerializer.extractSingle(env.store, MediocreVillain, json_hash);
 
   deepEqual(json, {
     "id": 1,
@@ -605,10 +208,10 @@ test("extractPolymorphic when the related data is not specified", function() {
 
 test("extractPolymorphic hasMany when the related data is not specified", function() {
   var json = {
-    popular_villain: {id: 1, name: "Dr Horrible"}
+    mediocre_villain: {id: 1, name: "Dr Horrible"}
   };
 
-  json = env.amsSerializer.extractSingle(env.store, PopularVillain, json);
+  json = env.amsSerializer.extractSingle(env.store, MediocreVillain, json);
 
   deepEqual(json, {
     "id": 1,
@@ -619,10 +222,10 @@ test("extractPolymorphic hasMany when the related data is not specified", functi
 
 test("extractPolymorphic does not break hasMany relationships", function() {
   var json = {
-    popular_villain: {id: 1, name: "Dr. Evil", evil_minions: []}
+    mediocre_villain: {id: 1, name: "Dr. Evil", evil_minions: []}
   };
 
-  json = env.amsSerializer.extractSingle(env.store, PopularVillain, json);
+  json = env.amsSerializer.extractSingle(env.store, MediocreVillain, json);
 
   deepEqual(json, {
     "id": 1,
